Revise tests to look for new behavior when marker is not found
Once upon a time, get_markers raised a ValueError if the marker_name
passed in doesn't match any existing marker name. Later, the API was
changed so that a warning was issued and None was returned if the
marker name was not found. In the most recent iteration of the API, an
empty table is to be returned with the proper column names.

It turns out three responses were being checked. In one case, the test
checked (and passed) if a warning was raised. In another, the test
passed if a ValueError was raised.

The setup for those cases remains but now the check is that the return
table has now rows and has the correct column names.